pyproject.toml
[tool.poetry]
name = "SensorPublisher"
version = "0.1.0"
description = ""
authors = ["kensan <9150185+kensan23@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.5"
boto3 = "^1.12.41"
python-dotenv = "^0.13.0"
bme680 = "^1.0.5"
apscheduler = "^3.6.3"
smbus = "^1.1.post2"
tenacity = "^6.2.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
autopep8 = "^1.5.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
